* Split actionCauseUtils tests in two for add/remove * Move Analytics util test toAnalyticsUtilsTest.ts Snow-347
